CounterfactualTraining

Documentation for CounterfactualTraining.

CounterfactualTraining.AbstractObjective

CounterfactualTraining.AdversarialObjective

CounterfactualTraining.AdversarialObjective

CounterfactualTraining.AdversarialObjective

CounterfactualTraining.EnergyDifferentialObjective

CounterfactualTraining.EnergyDifferentialObjective

CounterfactualTraining.EnergyDifferentialObjective

CounterfactualTraining.FullObjective

CounterfactualTraining.FullObjective

CounterfactualTraining.FullObjective

CounterfactualTraining.VanillaObjective

CounterfactualTraining.VanillaObjective

CounterfactualTraining.VanillaObjective

CounterfactualTraining.VanillaObjective

CounterfactualTraining.adv_loss

CounterfactualTraining.generate!

CounterfactualTraining.get_global_ae_criterium

CounterfactualTraining.get_last_valid_ae

CounterfactualTraining.implausibility

CounterfactualTraining.infer_domain_constraints

CounterfactualTraining.isvalid

CounterfactualTraining.protect_immutable!

CounterfactualTraining.reg_loss

CounterfactualTraining.setup_counterfactual_search

CounterfactualTraining.unwrap

CounterfactualTraining.AbstractObjective — Type

Base type of training objectives.

CounterfactualTraining.AdversarialObjective — Type

AdversarialObjective <: AbstractObjective

The AdversarialObjective is a concrete implementation of the AbstractObjective abstract type that optimizes for:

Standard classification objective (the discriminative task).

Adversarial classification objective on the counterfactuals (the explainability task).

CounterfactualTraining.AdversarialObjective — Type

(obj::AdversarialObjective)(
    yhat,
    y,
    energy_differential::Vector{<:AbstractFloat}=[0.0f0],
    regularization::Vector{<:AbstractFloat}=[0.0f0],
    adversarial_loss::Union{AbstractFloat,Vector{<:AbstractFloat}}=0.0f0;
    agg=mean,
    kwrgs...,
)

If the adversarial_loss has been computed already, obj::AdversarialObjective can be called directly on predictions yhat and labels y. The different loss components are then added together with a weighting vector lambda.

CounterfactualTraining.AdversarialObjective — Method

AdversarialObjective(;
    class_loss::Function=Flux.Losses.logitcrossentropy,
    lambda::Vector{<:AbstractFloat}=[1.0, 0.25]
)

Outer constructor for the AdversarialObjective type.

CounterfactualTraining.EnergyDifferentialObjective — Type

EnergyDifferentialObjective <: AbstractObjective

The EnergyDifferentialObjective is a concrete implementation of the AbstractObjective abstract type that optimizes for:

Standard classification objective (the discriminative task)

Energy differential between counterfactuals and observed data (the explainability task).

CounterfactualTraining.EnergyDifferentialObjective — Type

(obj::EnergyDifferentialObjective)(
    yhat,
    y,
    energy_differential::Vector{<:AbstractFloat}=[0.0f0],
    regularization::Vector{<:AbstractFloat}=[0.0f0],
    adversarial_loss::Union{AbstractFloat,Vector{<:AbstractFloat}}=0.0f0;
    agg=mean,
    kwrgs...,
)

If the energy_differential and regularization have been computed already, obj::EnergyDifferentialObjective can be called directly on predictions yhat and labels y. The different loss components are then added together with a weighting vector lambda.

CounterfactualTraining.EnergyDifferentialObjective — Method

EnergyDifferentialObjective(;
    class_loss::Function=Flux.Losses.logitcrossentropy, 
    lambda::Vector{<:AbstractFloat}=[1.0, 0.5, 0.1]
)

Outer constructor for the EnergyDifferentialObjective type.

CounterfactualTraining.FullObjective — Type

FullObjective <: AbstractObjective

The FullObjective is a concrete implementation of the AbstractObjective abstract type that optimizes for all three tasks:

Standard classification objective (the discriminative task)

Energy differential between counterfactuals and observed data (the explainability task).

Adversarial classification objective on the counterfactuals (the explainability task).

CounterfactualTraining.FullObjective — Type

(obj::FullObjective)(
    yhat,
    y,
    energy_differential::Vector{<:AbstractFloat}=[0.0f0],
    regularization::Vector{<:AbstractFloat}=[0.0f0],
    adversarial_loss::Union{AbstractFloat,Vector{<:AbstractFloat}}=[0.0f0];
    agg=mean,
    kwrgs...,
)

If the adversarial_loss has been computed already, obj::FullObjective can be called directly on predictions yhat and labels y. The different loss components are then added together with a weighting vector lambda.

CounterfactualTraining.FullObjective — Method

FullObjective(;
    class_loss::Function=Flux.Losses.logitcrossentropy,
    energy_differential::PenaltyOrFun=EnergyDifferential(),
    lambda::Vector{<:AbstractFloat}=[1.0, 0.5, 0.1, 0.25]
)

Outer constructor for the FullObjective type.

CounterfactualTraining.VanillaObjective — Type

(obj::VanillaObjective)(
    yhat,
    y,
    energy_differential::Vector{<:AbstractFloat}=[0.0f0],
    regularization::Vector{<:AbstractFloat}=[0.0f0],
    adversarial_loss::Union{AbstractFloat,Vector{<:AbstractFloat}}=0.0f0;
    agg=mean,
    kwrgs...,
)

obj::VanillaObjective can be called directly on predictions yhat and labels y.

CounterfactualTraining.VanillaObjective — Type

VanillaObjective <: AbstractObjective

The VanillaObjective is a concrete implementation of the AbstractObjective abstract type that optimizes for:

Standard classification objective (the discriminative task).

CounterfactualTraining.VanillaObjective — Method

VanillaObjective(class_loss, lambda; needs_ce=false)

Outer constructor to allow passing just class_loss and lambda as positional arguments.

CounterfactualTraining.VanillaObjective — Method

VanillaObjective(;
    class_loss::Function=Flux.Losses.logitcrossentropy,
    lambda::Vector{<:AbstractFloat}=[1.0],
    needs_ce::Bool=false,
)

Outer constructor for the VanillaObjective type.

CounterfactualTraining.adv_loss — Method

adv_loss(
    model, counterfactual, perturbations, targets; epsilon=2.0, p::Real=Inf, validities=nothing
)

Adversarial loss function.

CounterfactualTraining.generate! — Method

generate!(
    model,
    data,
    generator::AbstractGenerator;
    nsamples::Union{Int,Nothing}=nothing,
    convergence=Convergence.MaxIterConvergence(),
    parallelizer=nothing,
    input_encoder=nothing,
    verbose=1,
    domain=nothing,
)

This function generates counterfactual explanations for the whole dataset data or a subset thereof (nsamples). It is supposed to be used outside of the mini-batch training loop.

CounterfactualTraining.get_global_ae_criterium — Method

get_global_ae_criterium()

Get the global AE criterium.

CounterfactualTraining.get_last_valid_ae — Method

get_last_valid_ae(ce::CounterfactualExplanation)

A callback function used to store the last counterfactual that is also a valid adversarial example based on the global AE criterium (see get_global_ae_criterium).

CounterfactualTraining.implausibility — Method

implausibility(model, counterfactual, samples, targets)

Compute the implausibility (contrastive divergence) of the counterfactuals (counterfactual) with respect to samples in the target class. This is computed as the difference between negative logits indexed at the target class for the samples and the counterfactual.

CounterfactualTraining.infer_domain_constraints — Method

infer_domain_constraints(X::AbstractArray; nstd=3)

Automatically infers reasonable domain constraints for the counterfactuals.

CounterfactualTraining.isvalid — Method

isvalid(ce, model, data)

Checks if the label has been flipped.

CounterfactualTraining.protect_immutable! — Method

protect_immutable!(
    samples::AbstractArray,
    counterfactuals::AbstractArray,
    mutability::Union{Nothing,AbstractArray},
)

Protects immutable features from the contrastive divergence penalty.

CounterfactualTraining.reg_loss — Method

reg_loss(model, counterfactual, samples, targets)

Compute the regularization loss for the contrastice divergence.

CounterfactualTraining.setup_counterfactual_search — Method

setup_counterfactual_search(
    data,
    model,
    domain,
    input_encoder,
    mutability,
    nneighbours::Int64,
    nsamples::Union{Nothing,Int64},
)

Sets up the counterfactual search.

CounterfactualTraining.unwrap — Method

unwrap(train_set; labels=nothing)

Unwraps the data from a Flux.DataLoader or zip iterator. The output variables is assumed to be categorical. If no labels are provided, then 1 to n is used where n is the number of classes.